Move LedgerManager's request handling to LedgerCtx.
And messages to ledger_messages module to avoid import
cycle. NOTE: Still not working properly.
