+++ node/src/ledger/ledger_manager.rs
@@ -13,6 +13,7 @@ use std::sync::Arc;
 use std::thread;
 use std::time;
 
+use super::ledger_messages::{LedgerRequest, LedgerResponse};
 use super::ledger_service::LedgerCtx;
 use crate::block_producer::vrf_evaluator::{
     BlockProducerVrfEvaluatorLedgerService, DelegatorTable,
@@ -36,235 +37,6 @@ use mina_signer::CompressedPubKey;
 use openmina_node_account::AccountPublicKey;
 
 use crate::event_source::Event;
-
-/// The type enumerating different requests that can be made to the
-/// service. Each specific constructor has a specific response
-/// constructor associated with it. Unfortunately, this relationship
-/// can't be expressed in the Rust type system at the moment. For this
-/// reason this type is private while functions wrapping the whole call
-/// to the service are exposed as the service's methods.
-enum LedgerRequest {
-    AccountsSet {
-        snarked_ledger_hash: LedgerHash,
-        parent: LedgerAddress,
-        accounts: Vec<MinaBaseAccountBinableArgStableV2>,
-    }, // expected response: LedgerHash
-    BlockApply {
-        block: ArcBlockWithHash,
-        pred_block: ArcBlockWithHash,
-    }, // expected response: Success
-    ChildHashesGet {
-        snarked_ledger_hash: LedgerHash,
-        parent: LedgerAddress,
-    }, // expected response: ChildHashes
-    Commit {
-        ledgers_to_keep: BTreeSet<LedgerHash>,
-        root_snarked_ledger_updates: TransitionFrontierRootSnarkedLedgerUpdates,
-        needed_protocol_states: BTreeMap<StateHash, MinaStateProtocolStateValueStableV2>,
-        new_root: ArcBlockWithHash,
-        new_best_tip: ArcBlockWithHash,
-    }, // expected response: CommitResult
-    ComputeSnarkedLedgerHashes {
-        snarked_ledger_hash: LedgerHash,
-    }, // expected response: Success
-    CopySnarkedLedgerContentsForSync {
-        origin_snarked_ledger_hash: LedgerHash,
-        target_snarked_ledger_hash: LedgerHash,
-        overwrite: bool,
-    }, // expected response: SnarkedLedgerContentsCopied
-    GetProducerAndDelegates {
-        ledger_hash: LedgerHash,
-        producer: AccountPublicKey,
-    }, // expected response: ProducerAndDelegates
-    GetProducersWithDelegates {
-        ledger_hash: LedgerHash,
-        filter: fn(&CompressedPubKey) -> bool,
-    }, // expected response: ProducersWithDelegatesMap
-    GetMask {
-        ledger_hash: LedgerHash,
-    }, // expected response: LedgerMask
-    GetScanStateSummary {
-        ledger_hash: LedgerHash,
-    }, // expected response: ScanStateSummary
-    InsertGenesisLedger {
-        mask: Mask,
-    }, // expected response: Success
-    LedgerQuery {
-        ledger_hash: LedgerHash,
-        query: MinaLedgerSyncLedgerQueryStableV1,
-    }, // expected response: LedgerQueryResult
-    StagedLedgerAuxAndPendingCoinbase {
-        ledger_hash: LedgerHash,
-        protocol_states: BTreeMap<StateHash, MinaStateProtocolStateValueStableV2>,
-    }, // expected response: LedgerAuxAndCoinbaseResult
-    StagedLedgerDiffCreate {
-        pred_block: ArcBlockWithHash,
-        won_slot: BlockProducerWonSlot,
-        coinbase_receiver: NonZeroCurvePoint,
-        completed_snarks: BTreeMap<SnarkJobId, Snark>,
-        supercharge_coinbase: bool,
-    }, // expected response: StagedLedgerDiff
-    StagedLedgerReconstruct {
-        snarked_ledger_hash: LedgerHash,
-        parts: Option<Arc<StagedLedgerAuxAndPendingCoinbasesValid>>,
-    }, // expected response: Success
-    StakeProofSparseLedger {
-        staking_ledger: LedgerHash,
-        producer: NonZeroCurvePoint,
-        delegator: NonZeroCurvePoint,
-    }, // expected response: SparseLedgerBase
-}
-
-/// This type represents LedgerManager's responses to synchronous request.
-/// Each variant corresponds to a specific request type, but unfortunately,
-/// this relationship can't be expressed in the Rust type system at the moment.
-/// For this reason this type is kept private and its variants are converted to
-/// different types by their respective request handlers.
-#[derive(Debug)]
-enum LedgerResponse {
-    ChildHashes(LedgerHash, LedgerHash),
-    CommitResult(CommitResult),
-    LedgerAuxAndCoinbaseResult(Option<Arc<StagedLedgerAuxAndPendingCoinbases>>),
-    LedgerHash(LedgerHash),
-    LedgerMask(Option<(Mask, bool)>),
-    LedgerQueryResult(Option<MinaLedgerSyncLedgerAnswerStableV2>),
-    ProducerAndDelegates(DelegatorTable),
-    ProducersWithDelegatesMap(
-        Option<BTreeMap<AccountPublicKey, Vec<(ledger::AccountIndex, AccountPublicKey, u64)>>>,
-    ),
-    ScanStateSummary(Vec<Vec<RpcScanStateSummaryScanStateJob>>),
-    SnarkedLedgerContentsCopied(bool),
-    SparseLedgerBase(Option<MinaBaseSparseLedgerBaseStableV2>),
-    StagedLedgerDiff(StagedLedgerDiffCreateOutput),
-    Success, // operation was performed and result stored; nothing to return.
-}
-
-impl LedgerRequest {
-    fn handle(self, ledger_ctx: &mut LedgerCtx) -> Result<LedgerResponse, String> {
-        match self {
-            LedgerRequest::AccountsSet {
-                snarked_ledger_hash,
-                parent,
-                accounts,
-            } => ledger_ctx
-                .accounts_set(snarked_ledger_hash, &parent, accounts)
-                .map(LedgerResponse::LedgerHash),
-            LedgerRequest::BlockApply { block, pred_block } => ledger_ctx
-                .block_apply(block, pred_block)
-                .map(|()| LedgerResponse::Success),
-            LedgerRequest::ChildHashesGet {
-                snarked_ledger_hash,
-                parent,
-            } => ledger_ctx
-                .child_hashes_get(snarked_ledger_hash, &parent)
-                .map(|(left, right)| LedgerResponse::ChildHashes(left, right)),
-            LedgerRequest::Commit {
-                ledgers_to_keep,
-                root_snarked_ledger_updates,
-                needed_protocol_states,
-                new_root,
-                new_best_tip,
-            } => {
-                let res = ledger_ctx.commit(
-                    ledgers_to_keep,
-                    root_snarked_ledger_updates,
-                    needed_protocol_states,
-                    &new_root,
-                    &new_best_tip,
-                );
-                Ok(LedgerResponse::CommitResult(res))
-            }
-            LedgerRequest::ComputeSnarkedLedgerHashes {
-                snarked_ledger_hash,
-            } => ledger_ctx
-                .compute_snarked_ledger_hashes(&snarked_ledger_hash)
-                .map(|()| LedgerResponse::Success),
-            LedgerRequest::CopySnarkedLedgerContentsForSync {
-                origin_snarked_ledger_hash,
-                target_snarked_ledger_hash,
-                overwrite,
-            } => ledger_ctx
-                .copy_snarked_ledger_contents_for_sync(
-                    origin_snarked_ledger_hash,
-                    target_snarked_ledger_hash,
-                    overwrite,
-                )
-                .map(LedgerResponse::SnarkedLedgerContentsCopied),
-            LedgerRequest::GetMask { ledger_hash } => {
-                Ok(LedgerResponse::LedgerMask(ledger_ctx.mask(&ledger_hash)))
-            }
-            LedgerRequest::GetProducerAndDelegates {
-                ledger_hash,
-                producer,
-            } => {
-                let res = ledger_ctx.get_producer_and_delegates(ledger_hash, producer);
-                Ok(LedgerResponse::ProducerAndDelegates(res))
-            }
-            LedgerRequest::GetProducersWithDelegates {
-                ledger_hash,
-                filter,
-            } => {
-                let res = ledger_ctx.producers_with_delegates(&ledger_hash, filter);
-                Ok(LedgerResponse::ProducersWithDelegatesMap(res))
-            }
-            LedgerRequest::GetScanStateSummary { ledger_hash } => {
-                let res = ledger_ctx.scan_state_summary(ledger_hash);
-                Ok(LedgerResponse::ScanStateSummary(res))
-            }
-            LedgerRequest::InsertGenesisLedger { mask } => {
-                ledger_ctx.insert_genesis_ledger(mask);
-                Ok(LedgerResponse::Success)
-            }
-            LedgerRequest::LedgerQuery { ledger_hash, query } => {
-                let res = ledger_ctx.answer_ledger_query(ledger_hash, query);
-                Ok(LedgerResponse::LedgerQueryResult(res))
-            }
-            LedgerRequest::StagedLedgerAuxAndPendingCoinbase {
-                ledger_hash,
-                protocol_states,
-            } => {
-                let res =
-                    ledger_ctx.staged_ledger_aux_and_pending_coinbase(ledger_hash, protocol_states);
-                Ok(LedgerResponse::LedgerAuxAndCoinbaseResult(res))
-            }
-            LedgerRequest::StagedLedgerDiffCreate {
-                pred_block,
-                won_slot,
-                coinbase_receiver,
-                completed_snarks,
-                supercharge_coinbase,
-            } => ledger_ctx
-                .staged_ledger_diff_create(
-                    &pred_block,
-                    &won_slot,
-                    &coinbase_receiver,
-                    completed_snarks,
-                    supercharge_coinbase,
-                )
-                .map(LedgerResponse::StagedLedgerDiff),
-            LedgerRequest::StagedLedgerReconstruct {
-                snarked_ledger_hash,
-                parts,
-            } => {
-                ledger_ctx.spawn_staged_ledger_reconstruction(snarked_ledger_hash.clone(), parts);
-                openmina_core::info!(
-                    openmina_core::log::system_time();
-                    kind = "LedgerManager::LedgerRequest::handle",
-                    summary = format!("Spawned staged ledger reconstruction for {}.", snarked_ledger_hash)
-                );
-                Ok(LedgerResponse::Success)
-            }
-            LedgerRequest::StakeProofSparseLedger {
-                staking_ledger,
-                producer,
-                delegator,
-            } => {
-                let res = ledger_ctx.stake_proof_sparse_ledger(staking_ledger, producer, delegator);
-                Ok(LedgerResponse::SparseLedgerBase(res))
-            }
-        }
-    }
-}
 
 struct LedgerRequestWithChan {
     request: LedgerRequest,
@@ -296,7 +68,7 @@ impl LedgerManager {
                 ledger_ctx.staged_ledger_reconstructions_finalize();
                 match receiver.blocking_recv() {
                     Some(LedgerRequestWithChan { request, responder }) => {
-                        let result = request.handle(&mut ledger_ctx);
+                        let result = ledger_ctx.handle_request(request);
                         if let Some(r) = responder {
                             r.send(result).unwrap_or(())
                         }
